Moved the mock data in the pricechart component to a hook instead.
